VAE/requirements.txt
absl-py==0.7.0
astor==0.7.1
backcall==0.1.0
bleach==1.5.0
cloudpickle==0.8.0
colorama==0.4.1
cycler==0.10.0
dask==1.1.1
decorator==4.3.2
defusedxml==0.5.0
entrypoints==0.3
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
html5lib==0.9999999
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
mpmath==1.1.0
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.4
numpy==1.16.1
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.4
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.5.0
prompt-toolkit==2.0.8
protobuf==3.6.1
Pygments==2.3.1
pyparsing==2.3.1
python-barcode==0.9.0
python-dateutil==2.7.5
pytz==2018.9
PyWavelets==1.0.1
pywinpty==0.5.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
sympy==1.3
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.4.2
